feature(pagination): extract client-side pagination utilities
